+++ imports/parsers/getSongInfoYouTube.js
@@ -2,7 +2,7 @@
  * @author Thanh Tran
  */
 import { HTTP } from 'meteor/http';
-import { SongOrigin } from '../constants.js';
+import { SongOrigin, defaultThumbnailUrl } from '../constants.js';
 
 /**
  * Youtube URL parser module
@@ -10,29 +10,33 @@ import { SongOrigin } from '../constants.js';
  * @param  {[type]} songurl [description]
  * @return {[type]}         [description]
  */
-const getSongInfoYouTube = (songurl) => {
+const getSongInfoYouTube = songurl => {
 	const http = HTTP.call('GET', songurl);
 
+	const regex = /v=([^&]*)/;
+	let songId = regex.exec(songurl)[0];
+	songId = songId.split('=')[1];
+
 	const html = http.content;
-	// console.log(html);
 
 	if (html) {
 		console.log('Parsing the song data through html source');
 
-		const title = (/<meta.*name="title".*content="(.*?)">/ig).exec(html);
-		const thumbURL = (/watch-header[\s\S]*?yt-thumb-clip[\s\S]*?<img.*data-thumb="(.*?)"/igm).exec(html);
-		const user = (/watch-header[\s\S]*?"yt-user-info"\s*>[\s\S]*?>(.*)<\/a>/igm).exec(html);
-		let length = (/"length_seconds":"(\d*)"/).exec(html);
-		length = length ? (+length[1]) : 0;
+		const title = /<meta.*name="title".*content="(.*?)">/gi.exec(html);
+		// const thumbURL = /watch-header[\s\S]*?yt-thumb-clip[\s\S]*?<img.*data-thumb="(.*?)"/gim.exec(html);
+		const user = /watch-header[\s\S]*?"yt-user-info"\s*>[\s\S]*?>(.*)<\/a>/gim.exec(html);
+		let length = /"length_seconds":"(\d*)"/.exec(html);
+		const thumbURL = `https://img.youtube.com/vi/${songId}/0.jpg`;
+		length = length ? +length[1] : 0;
 		console.log(length);
 
 		if (!html.match(/"allow_embed":"1"/i)) {
 			return {
-				error: 'This YouTube video is not allowed to embed'
+				error: 'This YouTube video is not allowed to embed',
 			};
 		} else if (length > 10 * 60 || length < 10) {
 			return {
-				error: 'This YouTube video is too long (>10min) or too short (<10sec) to play'
+				error: 'This YouTube video is too long (>10min) or too short (<10sec) to play',
 			};
 		}
 
@@ -43,13 +47,13 @@ const getSongInfoYouTube = (songurl) => {
 			name: title[1],
 			artist: user[1], // this is not really means artist, this is the uploader field from youtube page
 			streamURL: songurl, // media element can play youtube URL directly
-			thumbURL: thumbURL[1],
-			play: 0
+			thumbURL: thumbURL || defaultThumbnailUrl,
+			play: 0,
 		};
 	}
 
 	return {
-		error: 'Can\'t parse and get song info from link'
+		error: "Can't parse and get song info from link",
 	};
 };